pyproject.toml
[project]
name = "reservation-restaurant"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "asgiref==3.9.1",
    "coverage==7.10.7",
    "django==5.2.5",
    "factory-boy==3.3.3",
    "faker==37.8.0",
    "jinja2==3.1.6",
    "markupsafe==3.0.2",
    "psycopg2-binary==2.9.10",
    "sqlparse==0.5.3",
    "tzdata==2025.2",
]
